pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spreads"
version = "0.1.0"
description = "Alpaca-based options spread scanner"
requires-python = ">=3.11"
dependencies = [
  "arq>=0.25.0",
  "alembic>=1.16.5",
  "fastapi>=0.118.0",
  "msgpack>=1.1.0",
  "pandas-market-calendars>=4.4.1",
  "python-dotenv>=1.1.1",
  "psycopg[binary]>=3.2.10",
  "py-vollib>=1.0.1",
  "rich>=13.7.1",
  "sqlalchemy>=2.0.43",
  "typer>=0.12.5",
  "uvicorn>=0.37.0",
  "websocket-client>=1.8.0",
  "websockets>=16.0",
  "pyyaml>=6.0.3",
]

[project.scripts]
spreads = "core.cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["packages/core", "packages/api"]
